Feature: add user preferences settings screen and navigation
